Testing Gearmotor with microswitch
Just added a case so that I can test how a gearmotor would work and see how I can use a microswitch to turn it on/off. Hoping that I can use the gearmotor instead of servo motors to angle the base of the bot. Ideally the gearmotor will spin about 90 degrees at which point it will hit a microswitch and stop.